refactor: 컴포넌트 네임스페이스 추가 (그룹화)
