+++ src/components/common/Badge.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import Badge from './Badge';
 
 const meta = {
-  title: 'Badge',
+  title: 'common/Badge',
   component: Badge,
 } satisfies Meta<typeof Badge>;
